manual reset of https://github.com/kolmafia/kolmafia/pull/11
