+++ src/net/sourceforge/kolmafia/textui/parsetree/Type.java
@@ -104,12 +104,6 @@ public class Type
 		return this.type == type;
 	}
 
-	@Override
-	public String toString()
-	{
-		return this.name;
-	}
-
 	public Type simpleType()
 	{
 		return this;
